MTA-812 moved lessons into a directory
