+++ resources/assets/js/app.js
@@ -66,7 +66,7 @@ Vue.component('invoiceCreate', require('./components/invoice/create/InvoiceCreat
 Vue.component('invoiceList', require('./components/invoice/list/InvoiceList.vue').default);
 Vue.component('listOfPayments', require('./components/invoice/payments/ListOfPayments.vue').default);
 Vue.component('holiday', require('./components/holiday/Holiday.vue').default);
-Vue.component('lessons', require('./components/Lessons.vue').default);
+Vue.component('lessons', require('./components/lessons/Lessons.vue').default);
 Vue.component('lessonCancelled', require('./components/lesson/Cancel.vue').default);
 Vue.component('lesson', require('./components/student/schedule/Lesson.vue').default);
 Vue.component('profile', require('./components/student/profile/Profile.vue').default);
