+++ resources/assets/js/components/student/schedule/lesson-template.html
@@ -70,10 +70,9 @@
                                        :disabled-dates='closed'
                                        :model-config="modelConfig">
                             <template v-slot="{ inputValue, inputEvents }">
-                                <input
-                                        class="px-2 py-1 border rounded focus:outline-none focus:border-blue-300 form-control"
-                                        :value="inputValue"
-                                        v-on="inputEvents"
+                                <input class="px-2 py-1 border rounded focus:outline-none focus:border-blue-300 form-control"
+                                       :value="inputValue"
+                                       v-on="inputEvents"
                                 />
                             </template>
                         </v-date-picker>
